+++ spec/config_specs.clj
@@ -1,11 +1,9 @@
 (ns config-specs
-  (:use io.aviso.config
+  (:use sharetribe.config
         speclj.core)
   (:require [com.stuartsierra.component :as component]
-            [clojure.future :refer [int?
-                                    simple-keyword?]]
-            [clojure.spec :as s]
-            [clojure.spec.test :as stest])
+            [clojure.spec.alpha :as s]
+            [clojure.spec.test.alpha :as stest])
   (:import (clojure.lang ExceptionInfo)))